\name{RIVSparse-package}
\alias{RIVSparse-package}
\alias{RIVSparse}
\docType{package}
\title{\packageTitle{RIVSparse}}
\description{\packageDescription{RIVSparse}}
\section{Package Content}{\packageIndices{RIVSparse}}
\author{\packageAuthor{RIVSparse}}
\section{Maintainer}{\packageMaintainer{RIVSparse}}
\keyword{package}
